requirements.txt
# YouTube Analytics Multi-Agent System - Consolidated Requirements

# Core AI/Agent Dependencies
google-adk
google-genai
google-generativeai
openai

# YouTube & Google APIs
google-api-python-client
yt-dlp

# Data Processing & Analysis
pandas
numpy
scipy
tiktoken

# Visualization
matplotlib
seaborn

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Utilities
python-dotenv>=1.0.0
